Avoid interning Org tag string faces
The `intern' function, when called on a string which has
text-properties, interns those text properties, so calling `symbol-name'
on the symbol returns a string with the same properties.  This results
in the interned tag symbols having the org-tag face, which is awkward if
those strings are later inserted into a buffer.

Fixes #35.